moved the view component into the component folder
