+++ src/app/components/view/view.component.spec.ts
